# Swagger Documentation.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "My API",
    "version": "v1"
  },
  "paths": {
    "/Administration/register": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Administration/logon": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/LogonModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogonModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LogonModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LogonModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Administration/createRole": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/RoleViev"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleViev"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleViev"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoleViev"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Administration/provideUserToRole": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Administration/takeUserFromRole": {
      "post": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ManipWithUserRole"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Administration/DeleteUser": {
      "delete": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Administration/GetUser": {
      "get": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/Administration/DeleteRole": {
      "delete": {
        "tags": [
          "Administration"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Administration/Update": {
      "put": {
        "tags": [
          "Administration"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LogonModel": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 16,
            "minLength": 4,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ManipWithUserRole": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "maxLength": 20,
            "minLength": 2,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RegisterModel": {
        "required": [
          "email",
          "firstName",
          "lastName",
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "maxLength": 16,
            "minLength": 4,
            "type": "string"
          },
          "passwordConfirm": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          },
          "phoneNumber": {
            "type": "string",
            "format": "tel",
            "nullable": true
          },
          "login": {
            "maxLength": 20,
            "minLength": 4,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "RoleViev": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}